implement get tarif by name and impl exception handler

+++ src/main/java/com/marian/tennis/api/tarifs/entity/TarifsEntity.java
@@ -32,7 +32,7 @@ public class TarifsEntity {
     private String name;
     @NotNull
     @Column(name = "prix")
-    private double prix;
+    private float prix;
     @Column(name = "start_date")
     @NotNull
     @DateTimeFormat(pattern = "yyyy-MM-dd")
